+++ lib/generators/templates/create_cache_database.rb
@@ -8,28 +8,5 @@ class Migration < ::ActiveRecord::Migration[5.2]
       t.timestamp :created_at, null: false, index: true
       t.timestamp :expires_at, index: true
     end
-
-    if postgresql?
-      # Unlogged tables come with at least 50% write performance increase, but come with multiple downsides you need to be aware off:
-      # - No validation, data will be lost in case of forced restart.
-      # - No replication to read replicas
-      #
-      # Since we're working with cache here, these drawbacks are accaptable. But you can turn that off, just comment next line out.
-      ActiveRecord::Base.connection.execute("ALTER TABLE activesupport_cache_entries SET UNLOGGED")
-    end
-  end
-
-  private
-
-  def adapter
-    if ActiveRecord::VERSION::STRING.to_f >= 6.1
-      ActiveRecord::Base.connection_db_config.adapter.to_s
-    else
-      ActiveRecord::Base.connection_config[:adapter].to_s
-    end
-  end
-
-  def postgresql?
-    adapter =~ /postg/
   end
 end
